+++ lab03/decorator/Stream/ReplaceEncryptor.cpp
@@ -6,7 +6,7 @@ ReplaceEncryptor::ReplaceEncryptor(size_t key)
 	, decodeArr(encodeArr.size())
 {
 	std::iota(encodeArr.begin(), encodeArr.end(), '\0');
-	std::shuffle(encodeArr.begin(), encodeArr.end(), std::mt19937(key));
+	std::shuffle(encodeArr.begin(), encodeArr.end(), std::mt19937(static_cast<uint32_t>(key)));
 	for (size_t i = 0; i < encodeArr.size(); ++i)
 	{
 		decodeArr[encodeArr[i]] = static_cast<uint8_t>(i);
